using int instead of long for proxy lifetime

+++ FrameworkSystem/Client/ProxyManagerClient.py
@@ -1,6 +1,5 @@
 """ ProxyManagementAPI has the functions to "talk" to the ProxyManagement service
 """
-from past.builtins import long
 import six
 import os
 import datetime
@@ -241,10 +240,10 @@ class ProxyManagerClient(object):
       rpcClient = RPCClient("Framework/ProxyManager", timeout=120)
     if token:
       retVal = rpcClient.getProxyWithToken(userDN, userGroup, req.dumpRequest()['Value'],
-                                           long(cacheTime + requiredTimeLeft), token)
+                                           int(cacheTime + requiredTimeLeft), token)
     else:
       retVal = rpcClient.getProxy(userDN, userGroup, req.dumpRequest()['Value'],
-                                  long(cacheTime + requiredTimeLeft))
+                                  int(cacheTime + requiredTimeLeft))
     if not retVal['OK']:
       return retVal
     chain = X509Chain(keyObj=req.getPKey())
@@ -307,11 +306,11 @@ class ProxyManagerClient(object):
       rpcClient = RPCClient("Framework/ProxyManager", timeout=120)
     if token:
       retVal = rpcClient.getVOMSProxyWithToken(userDN, userGroup, req.dumpRequest()['Value'],
-                                               long(cacheTime + requiredTimeLeft), token, requiredVOMSAttribute)
+                                               int(cacheTime + requiredTimeLeft), token, requiredVOMSAttribute)
 
     else:
       retVal = rpcClient.getVOMSProxy(userDN, userGroup, req.dumpRequest()['Value'],
-                                      long(cacheTime + requiredTimeLeft), requiredVOMSAttribute)
+                                      int(cacheTime + requiredTimeLeft), requiredVOMSAttribute)
     if not retVal['OK']:
       return retVal
     chain = X509Chain(keyObj=req.getPKey())
